Single threaded task executor
